requirements.txt
python-telegram-bot==20.7
requests==2.31.0
python-dotenv==1.0.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
lxml==4.9.3
